Hill climbing. It works!
The new values are written to new value .txt files named:
        "service_climbed_values.txt"
They are stored in the directory:
    "pal/heuristics/hill_climb/values_for_hill_climb"
If the hill climb is terminated prematurely, make sure to run tests on
the climbed values, as there is no guarantee that they are better than
the previous magic numbers.

There are a bunch of constants (read: magic numbers) at the top of the file.
They are algorithmic voodoo. Quinn and I think that they seem reasonable,
but that by no means mean that they are. If anyone wants to fiddle with them
they should.